scripts/rebuld-and-deploy: add a --no-dl option
This option allow to tweak the source code for Neon Beat command and
still be able to deploy it easily (ie without having to push it to
github)

+++ scripts/rebuild-and-deploy.sh
@@ -12,9 +12,20 @@ set -e
 SCRIPT_DIR=$(dirname "$(realpath $0)")
 TOP_DIR=${SCRIPT_DIR}/..
 
-rm -rf ${TOP_DIR}/dl/neon-beat-*
-make -C ${TOP_DIR} neon-beat-{{game,admin}-frontend,backend}-dirclean all
-make -C ${TOP_DIR} neon-beat-{{game,admin}-frontend,backend}
+DL=1
+if [ "$1" == "--no-dl" ]
+then
+	DL=0
+fi
+
+if [ ${DL} -eq 1 ]
+then
+	rm -rf ${TOP_DIR}/dl/neon-beat-*
+	make -C ${TOP_DIR} neon-beat-{{game,admin}-frontend,backend}-dirclean all
+	make -C ${TOP_DIR} neon-beat-{{game,admin}-frontend,backend}
+else
+	make -C ${TOP_DIR} neon-beat-{{game,admin}-frontend,backend}-rebuild
+fi
 ssh nbc "systemctl stop neon-beat-backend"
 scp -O ${TOP_DIR}/output/target/usr/bin/neon-beat-back nbc:/usr/bin
 scp -rO ${TOP_DIR}/output/target/srv/www nbc:/srv
